requirements.txt
accelerate==1.2.1
beautifulsoup4==4.12.3
blobfile==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
nautilus-open-any-terminal==0.6.0
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
